src/python-data-analysis/requirements.txt
ruff # formatter, linter
mypy # Type annoataion linter

psutil
pySerial 

ipykernel

matplotlib
pandas